Split stack experience details declaration into dedicated header
